feat: refactorización del github de vista; y de la vista welcome
