+++ exam-craft-extension-frontend/src/css/CommonText.css
@@ -1,0 +1,92 @@
+.main-content {
+  flex: 1;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: flex-start;
+  padding: 80px 20px;
+  gap: 20px;
+  text-align: center;
+  background-color: #EFDECD; 
+  min-height: 100vh;
+  height: 100vh;        
+  overflow: hidden;     
+}
+
+.main-title {
+  color: #5D3A3A; 
+  font-size: 3rem;
+  font-weight: 900;
+  text-transform: uppercase;
+  text-align: center;
+  font-style: italic;
+  margin-bottom: 15px;
+  letter-spacing: 1px;
+  text-shadow: 2px 2px 0px rgba(255,255,255,0.4);
+}
+
+.subtitle-badge {
+  background-color: #E6DFF0; 
+  color: #6A4C93;  
+  padding: 10px 40px;
+  border-radius: 50px;
+  font-weight: 600;
+  font-size: 1.3rem;
+  margin-bottom: 50px;
+  box-shadow: 0 4px 15px rgba(155, 114, 207, 0.25);
+  border: none;
+  border-radius: 50px;      
+  padding: 10px 40px;
+  font-family: 'Segoe UI', 'Roboto', sans-serif;
+  letter-spacing: 0.5px;
+}
+
+.btn-back {
+  background-color: #3E2723; 
+  color: #FDF5E6;            
+  border: none;
+  padding: 14px 35px;
+  border-radius: 50px;      
+  font-weight: bold;
+  font-size: 16px;
+  cursor: pointer;
+  box-shadow: 0 4px 10px rgba(0,0,0,0.3); 
+  z-index: 100;              
+  transition: transform 0.2s, background-color 0.2s;
+}
+
+.btn-back:hover { 
+  transform: scale(1.1); 
+  background-color: #5D3A3A; 
+}
+
+.btn-delete {
+    position: absolute;
+    top: -10px;
+    right: -10px;
+    background-color: #ff4d4f;
+    color: white;
+    border: none;
+    border-radius: 50%;
+    width: 28px;
+    height: 28px;
+    font-weight: bold;
+    font-size: 16px;
+    cursor: pointer;
+    z-index: 10;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+}
+
+.btn-icon {
+    cursor: pointer;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    height: 110px;
+    width: 100%;
+    background: none;
+    border: none;
+    padding: 0;
+}
