+++ exam-craft-extension-frontend/src/css/Comun.css
@@ -1,95 +1,9 @@
-/* =========================================
-   RESET Y BASE
-   ========================================= */
-* {
-  box-sizing: border-box;
-  margin: 0;
-  padding: 0;
-  font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
-}
 
-.exam-app {
-  display: flex;
-  flex-direction: column;
-  height: 100vh;
-  background-color: #EFDECD; 
-  overflow: hidden;
-}
 
-/* =========================================
-   HEADER 
-   ========================================= */
-.app-header {
-  background-color: #B48574; 
-  padding: 15px 30px;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  box-shadow: 0 2px 5px rgba(0,0,0,0.1);
-  height: 70px;
-  flex-shrink: 0;
-}
-
-.header-left {
-  display: flex;
-  align-items: center;
-  gap: 15px;
-  font-weight: bold;
-  color: #3E2723;
-  font-size: 1.2rem;
-}
-
-.header-right {
-  color: #3E2723;
-  font-weight: bold;
-  font-size: 1rem;
-}
-
-.logo-icon {
-  display: flex; align-items: center; cursor: pointer;
-}
 
 /* =========================================
    MAIN CONTENT
    ========================================= */
-.main-content {
-  flex: 1;
-  display: flex;
-  flex-direction: column;
-  align-items: center;     
-  justify-content: center; 
-  gap: 20px;
-  padding: 20px;
-  overflow: hidden; 
-}
-
-.main-title {
-  font-size: 3rem;
-  color: #5D3A3A; 
-  text-transform: uppercase;
-  font-style: italic;
-  font-weight: 900;
-  text-shadow: 2px 2px 0px rgba(255,255,255,0.4);
-  letter-spacing: 1px;
-  margin-bottom: 10px;
-  text-align: center;
-}
-
-.main-title.small { font-size: 2rem; }
-
-.subtitle-badge {
-  background-color: #E6DFF0; 
-  color: #6A4C93;            
-  border: none;
-  border-radius: 50px;      
-  padding: 10px 40px;
-  margin-bottom: 40px;
-  box-shadow: 0 4px 15px rgba(155, 114, 207, 0.25);
-  font-family: 'Segoe UI', 'Roboto', sans-serif;
-  font-weight: 600;
-  font-size: 1.3rem;
-  letter-spacing: 0.5px;
-}
 
 /* =========================================
    HOME CARDS
@@ -145,45 +59,6 @@
 /* =========================================
    GITHUB SCREEN
    ========================================= */
-.content-card {
-  background-color: rgba(255, 255, 255, 0.6);
-  backdrop-filter: blur(10px);
-  border-radius: 20px;
-  padding: 30px;
-  width: 90%;
-  max-width: 600px; 
-  box-shadow: 0 4px 15px rgba(62, 39, 35, 0.1);
-  display: flex;
-  flex-direction: column;
-  gap: 20px;
-  margin: 0 auto; 
-}
-
-.btn-back {
-  background-color: #3E2723; 
-  color: #FDF5E6;            
-  border: none;
-  padding: 10px 20px;
-  border-radius: 50px;      
-  font-weight: bold;
-  cursor: pointer;
-  box-shadow: 0 4px 10px rgba(0,0,0,0.3); 
-  z-index: 100;              
-  transition: transform 0.2s, background-color 0.2s;
-}
-
-.btn-back:hover { transform: scale(1.1); background-color: #5D3A3A; }
-
-.profiles-grid { display: grid; grid-template-columns: 1fr 1fr; gap: 15px; }
-.user-card { background-color: #fff; border-radius: 15px; padding: 15px; display: flex; align-items: center; gap: 10px; border: 2px solid #A67C6D; }
-.avatar { border-radius: 50%; border: 2px solid #B48574; }
-.repo-card { background-color: #A67C6D; color: white; padding: 20px; border-radius: 15px; text-align: center; }
-.repo-link { display: inline-block; background-color: #FDF5E6; color: #3E2723; padding: 8px 20px; border-radius: 20px; text-decoration: none; font-weight: bold; }
-.btn-floating-github {
-  position: absolute; bottom: 20px; right: 20px; 
-  background-color: #3E2723; color: #FDF5E6; border: none; padding: 10px 20px;
-  border-radius: 50px; font-weight: bold; cursor: pointer; box-shadow: 0 4px 10px rgba(0,0,0,0.3); z-index: 100; transition: transform 0.2s;
-}
 
 /* =========================================
    WORKFLOW SPECIFIC STYLES
@@ -397,35 +272,4 @@
 .menu-btn1:active {
   transform: translateY(0); 
   box-shadow: 0 2px 4px rgba(0,0,0,0.2);
-}
-
-/* =========================================
-   BREADCRUMBS STYLES
-   ========================================= */
-.breadcrumb-nav {
-  display: flex;
-  align-items: center;
-  gap: 4px;
-  font-size: 18px; 
-}
-
-.breadcrumb-link {
-  cursor: pointer;
-  transition: opacity 0.2s;
-  font-weight: 500;
-}
-
-.breadcrumb-link:hover {
-  opacity: 0.7;
-  text-decoration: underline;
-}
-
-.breadcrumb-separator {
-  color: #4e3506; 
-  margin: 0 2px;
-}
-
-.breadcrumb-current {
-  color: #210e01; 
-  font-weight: bold;
 }
