+++ exam-craft-extension-frontend/src/screens/principal/css/GitHub.css
@@ -1,0 +1,131 @@
+.btn-floating-github {
+  position: fixed;
+  bottom: 50px;                         
+  right: 30px;
+  background-color: #3E2723;
+  color: #FDF5E6;
+  border: none;
+  padding: 16px 30px;
+  border-radius: 50px;
+  font-weight: bold;
+  font-size: 16px;
+  cursor: pointer;
+  box-shadow: 0 4px 10px rgba(0,0,0,0.3);
+  transition: transform 0.2s;
+}
+
+.btn-floating-github:hover {
+  transform: scale(1.05);
+}
+
+.content-card-github {
+  background-color: rgba(255, 255, 255, 0.6);
+  backdrop-filter: blur(10px);
+  border-radius: 20px;
+  padding: 40px;
+  width: 90%;
+  max-width: 900px; 
+  box-shadow: 0 4px 15px rgba(62, 39, 35, 0.1);
+  display: flex;
+  flex-direction: column;
+  gap: 20px;
+  margin: 0 auto;
+}
+
+.profiles-grid { 
+  display: grid; 
+  grid-template-columns: 1fr 1fr; 
+  gap: 20px; 
+  align-items: stretch;
+}
+
+.user-card {
+  background-color: #fff;
+  border-radius: 15px;
+  padding: 25px;
+  display: flex;
+  align-items: center;
+  gap: 15px;
+  border: 2px solid #A67C6D;
+  min-height: 100px;  
+  height: 100%;
+}
+
+.user-details {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: flex-start;
+  gap: 6px;
+  min-width: 150px;
+}
+
+.user-name {
+  font-size: 20px;
+  font-weight: 700;
+  color: #3E2723;
+  margin: 0;
+  padding: 0;
+  line-height: 1;
+  display: block;
+}
+
+.user-repos {
+  font-size: 15px;
+  color: #7a5c4f;
+  margin: 0;
+  padding: 0;
+  line-height: 1;
+  display: block;
+}
+
+.avatar {
+  border-radius: 50%;
+  width: 80px;
+  height: 80px;
+  border: 2px solid;
+  object-fit: cover;  
+  flex-shrink: 0;   
+}
+
+.repo-card { 
+  background-color: #A67C6D; 
+  color: white; 
+  padding: 20px;   
+  border-radius: 15px; 
+  text-align: center;
+  font-size: 16px; 
+}
+
+.repo-card h3 {
+  font-size: 22px;
+  font-weight: 800;
+  margin: 0 0 12px 0;
+}
+
+.repo-name {
+  font-size: 18px;
+  font-weight: 600;
+  margin-bottom: 8px;
+}
+
+.repo-desc {
+  font-size: 16px;
+  margin-bottom: 8px;
+  opacity: 0.9;
+}
+
+.repo-stars {
+  font-size: 18px;
+  margin-bottom: 16px;
+}
+
+.repo-link { 
+  display: inline-block; 
+  background-color: #FDF5E6; 
+  color: #3E2723; 
+  padding: 8px 20px; 
+  border-radius: 20px; 
+  text-decoration: none; 
+  font-weight: bold; 
+}
