+++ exam-craft-extension-frontend/src/tabs/index.tsx
@@ -1,13 +1,12 @@
 import { useState } from "react"
-import WelcomeScreen from "../screens/WelcomeScreen"
-import GithubScreen from "../screens/GithubScreen"
-import CreateExamScreen from "../screens/CreateExamScreen"
-import CreateExamByPartsScreen from "../screens/CreateExamByPartsScreen"
+import WelcomeScreen from "../screens/principal/WelcomeScreen"
+import GithubScreen from "../screens/principal/GithubScreen"
+import CreateExamScreen from "../screens/principal/CreateExamScreen"
+import CreateExamByPartsScreen from "../screens/principal/CreateExamByPartsScreen"
 import DomainSelectionScreen from "../screens/examStatementGeneration/DomainSelectionScreen"
 import DomainWorkflowScreen from "../screens/examStatementGeneration/DomainWorkflowScreen"
 import DiagramUMLScreen from "../screens/examStatementGeneration/DiagramUMLScreen"
 import FinishFunctionalExtensionScreen from "../screens/examStatementGeneration/FinishFunctionalExtensionScreen"
-import "/assets/main.css"
 import AttributesConstraintsWorkflowScreen from "~src/screens/examStatementGeneration/AttributesConstraintsWorkflowScreen"
 import GenerationTestAtributesScreen from "../screens/codeGeneration/GenerationTestAtributesScreen"
 import GeneralGenerationTestScreen from "../screens/codeGeneration/exam/GeneralGenerationTestScreen"
